+++ Entidades/DetalleCompra.cs
@@ -35,8 +35,8 @@ namespace Entidades
             }
         }
 
-        private decimal _PrecioUnidad;
-        public decimal PrecioUnidad
+        private double _PrecioUnidad;
+        public double PrecioUnidad
         {
             get
             {
@@ -107,7 +107,7 @@ namespace Entidades
         public DetalleCompra() { }
 
         public DetalleCompra(int idDetalles,
-                              decimal precioUnidad,
+                              double precioUnidad,
                               int cantidad,
                               int iva,
                               OrdenCompra ordenCompra,
